W: need to finish makefile

+++ srcs/requirements/wordpress/tools/init-only-if-mariadb-is-up.sh
@@ -1,19 +1,20 @@
-#! /bin/bash
+#!/bin/bash
 
-rm -fr /var/www/html/wp-config-sample.php \
-  && \
-  wp core install --allow-root \
-  --path=/var/www/html \
-  --title="42Project InceptionHELL" \
-  --url=${SERVER_NAME} \
-  --admin_user=${ROOT_USER} \
-  --admin_password=${ROOT_PW} \
-  --admin_email=${ROOT_MAIL} \
-  && \
-  wp user create --allow-root \
-  --path=/var/www/html \
-  "${WP_USER}" "${WP_MAIL}" \
-  --user_pass=${WP_PW} \
-  --role="author"
+if [ -f /var/www/html/wp-config-sample.php ]; then
+	rm -rf /var/www/html/wp-config-sample.php
+	wp core install --allow-root \
+		--path=/var/www/html \
+		--title="42SP InceptionHELL" \
+		--url=$SERVER_NAME \
+		--admin_user=$ROOT_USER \
+		--admin_password=$ROOT_PW \
+		--admin_email=$ROOT_MAIL
 
-/usr/bin/php-fpm -F
+	wp user create --allow-root	\
+		--path=/var/www/html \
+		"$WP_USER" "$WP_MAIL" \
+		--user_pass=$WP_PW \
+		--role='author'
+fi
+
+exec php-fpm7.4 -F
